+++ bindings/nostr-sdk-js/src/client/signer/nip46.rs
@@ -5,7 +5,7 @@
 use core::ops::Deref;
 
 use js_sys::Array;
-use nostr_sdk::signer::Nip46Signer;
+use nostr_connect::prelude::*;
 use wasm_bindgen::prelude::*;
 
 use crate::duration::JsDuration;
@@ -14,36 +14,36 @@ use crate::protocol::key::{JsKeys, JsPublicKey};
 use crate::protocol::nips::nip46::JsNostrConnectURI;
 use crate::JsStringArray;
 
-#[wasm_bindgen(js_name = Nip46Signer)]
-pub struct JsNip46Signer {
-    inner: Nip46Signer,
+#[wasm_bindgen(js_name = NostrConnect)]
+pub struct JsNostrConnect {
+    inner: NostrConnect,
 }
 
-impl Deref for JsNip46Signer {
-    type Target = Nip46Signer;
+impl Deref for JsNostrConnect {
+    type Target = NostrConnect;
 
     fn deref(&self) -> &Self::Target {
         &self.inner
     }
 }
 
-impl From<Nip46Signer> for JsNip46Signer {
-    fn from(inner: Nip46Signer) -> Self {
+impl From<NostrConnect> for JsNostrConnect {
+    fn from(inner: NostrConnect) -> Self {
         Self { inner }
     }
 }
 
-#[wasm_bindgen(js_class = Nip46Signer)]
-impl JsNip46Signer {
+#[wasm_bindgen(js_class = NostrConnect)]
+impl JsNostrConnect {
     /// Construct Nostr Connect client
     #[wasm_bindgen]
     pub fn new(
         uri: &JsNostrConnectURI,
         app_keys: &JsKeys,
         timeout: &JsDuration,
-    ) -> Result<JsNip46Signer> {
+    ) -> Result<JsNostrConnect> {
         Ok(Self {
-            inner: Nip46Signer::new(
+            inner: NostrConnect::new(
                 uri.deref().clone(),
                 app_keys.deref().clone(),
                 **timeout,